1箇所でしか使われていないので、tipConfirmedの中身をTipコンポーネントへばらす
